backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
yfinance==0.2.40
httpx==0.28.1
pandas==2.2.2
openpyxl==3.1.2
python-multipart==0.0.9
pydantic==2.12.5
pydantic-settings==2.2.1
google-genai==1.69.0
weasyprint==62.3
apscheduler==3.10.4
sendgrid==6.11.0
redis==5.0.4
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
nsetools==1.0.11
